+++ C-Interop/01-Hello-World/project/src/hello-world.c
@@ -1,0 +1,6 @@
+#include <stdio.h>
+
+void hello_world()
+{
+	printf("Hello world from C!\n");
+}