+++ C-Interop/01-Hello-World/project/src/main.swift
@@ -1,0 +1,2 @@
+print("Hello world from swift!")
+hello_world()